requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
python-nmap
requests
tqdm
uuid
python-whois
prettytable
argparse
cargo-lite
paramiko
fire
tenacity
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3 ; sys_platform == 'win32'
dnspython==2.0.0
exrex==0.10.5
fire==0.3.1
future==0.18.2
idna==2.10
loguru==0.5.2
pysocks==1.7.1
requests-file==1.5.1
requests[socks]==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlalchemy==1.3.19
tenacity==6.2.0
termcolor==1.1.0
tldextract==2.2.3
tqdm==4.48.2
treelib==1.6.1
urllib3==1.25.10
win32-setctime==1.0.2 ; sys_platform == 'win32'
simplejson